% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-inflation.R
\docType{data}
\name{annual_inflation_CPI}
\alias{annual_inflation_CPI}
\title{Consumer Price Inflation (CPI)}
\format{data.frame with a row for each year. Variables are:
\describe{
   \item{year}{to year}
   \item{annual_change}{percentage change from previous year}
   \item{prop_change_from_prev_year}{Proportion change from previous year}
}}
\source{
\url{https://www.ons.gov.uk/economy/inflationandpriceindices/datasets/consumerpriceinflation}
}
\usage{
annual_inflation_CPI
}
\description{
Consumer Price Inflation (CPI)
}
\keyword{datasets}
