% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-inflation.R
\docType{data}
\name{annual_inflation_HCHS}
\alias{annual_inflation_HCHS}
\title{PSSRU annual inflation hospital and community health services (HCHS)}
\format{data.frame with a row for each year. From-to July. Variables are:
\describe{
   \item{year}{to year}
   \item{HCHS_pay}{Pecentage change from previous year}
   \item{prop_change_from_prev_year}{Proportion change from previous year}
}}
\source{
2017 \url{https://www.pssru.ac.uk/pub/uc/uc2017/sources-of-information.pdf}
}
\usage{
annual_inflation_HCHS
}
\description{
Until 2016/2017, a hospital & community health services (HCHS) index was calculated
by the Department of Health (now DHSC). The hospital and community health services (HCHS)
pay and price inflation was a weighted average of two separate inflation indices: the pay
cost index was calculated using the annual increase in NHS salaries and the Health Service
Cost Index (HSCI) measured the price change for each of 40 sub-indices of goods and
services purchased by the HCHS. These were weighted according to the proportion of
expenditure on pay and prices to give the HCHS pay and prices index.
In 2016, this index was discontinued, and in 2018, in collaboration with the DHSC,
PSSRU have developed a new inflation index.
}
\keyword{datasets}
