% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_QALY_pop.R
\name{calc_QALY}
\alias{calc_QALY}
\title{Calculate Quality-Adjusted Life Years}
\usage{
calc_QALY(utility = NA, intervals = NA, age = NA, start_delay = 0,
  discount_rate = 0.035, utility_method = "add")
}
\arguments{
\item{utility}{Vector of values between 0 and 1 (1 - utility loss)}

\item{intervals}{Time intervals for each utility}

\item{age}{Year of age}

\item{start_delay}{What time delay to origin, to shift discounting}

\item{discount_rate}{default 3.5\%}

\item{utility_method}{Should the yearly QALYs be summed to a scalar? \code{add} or \code{prod}}
}
\description{
Discounted total QALYs upto a defined time horizon.
This is a simpler function.
An alternative Method is also available (see \code{\link{total_QALYs}}).
}
\details{
Uses the following formula, for year  \code{i}:

\deqn{ \sum prop_year(i) * utility(i) * QoL(age(i)) * discount_factor(i) }

for  \code{i} = 1, ..., \code{time_horizon}.

\code{prop_year} is useful for fractions of years at the start and end of the period.
However, since we may not know this then may not be necessary.
}
\examples{
calc_QALY(utility = 0.9,
          age = 13,
          intervals = 49)

}
\references{
Sassi, Franco, Health Policy and Planning, 5, 402-408, Calculating QALYs,
comparing QALY and DALY calculations, volume 21, 2006
}
\seealso{
\code{\link{calc_QALY_CFR}},
         \code{\link{calc_QALY_population}}
         \code{\link{total_QALYs}}
}
