% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_QALY_pop.R
\name{calc_QALY_population}
\alias{calc_QALY_population}
\title{Calculate QALYs for population}
\usage{
calc_QALY_population(utility, intervals = NA, age = NA,
  start_delay = NA, discount_rate = 0.035, sum_res = TRUE, ...)
}
\arguments{
\item{utility}{Vector of utilities for each year in to the future, between 0 and 1}

\item{intervals}{Time intervals for each utility}

\item{age}{Vector of ages at start}

\item{start_delay}{What time delay to origin, to shift discounting}

\item{discount_rate}{default 3.5\%}

\item{sum_res}{Should the yearly QALYs be summed to a scalar?}

\item{...}{Additional arguments}
}
\value{
QALY vector
}
\description{
This is a wrapper function for \code{calc_QALY} over
multiple time horizons (e.g. individuals).
}
\details{
Assume that the utilities are the same for all individuals.
}
\examples{

}
\seealso{
\code{\link{calc_QALY_CFR}},
         \code{\link{calc_QALY}}
}
