% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discount.R
\name{discount}
\alias{discount}
\title{Discounted Values Over Time}
\usage{
discount(discount_rate = 0.035, t_limit = 100)
}
\arguments{
\item{discount_rate}{Discount factor, default at 3.5\%}

\item{t_limit}{Time period (positive integer) to discount over starting from 1}
}
\value{
Discounted value for each time point up to \code{t_limit}
}
\description{
Dscounted value, e.g. cost or health (QALY), for each time point, e.g. year.
E.g. a QALY in the future is worth less to us now because of 'interest'
or conversely we'd need more QALYs now to have a QALY further in the future.
}
\details{
Formula used is
 \deqn{ 1/(1 + discount_rate)^#years}
}
\examples{

D <- discount(t_limit = 10)
utility <- 0.97

# Discounted QALYs upto 10 years
sum(utility * D)

}
\references{
Severens, Johan L and Milne, Richard J,
Value in Health, 4, Discounting Health Outcomes in Economic Evaluation : The Ongoing Debate,
volume 7, 2004
}
