% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillin_utilities.R
\name{fillin_trailing_utilities}
\alias{fillin_trailing_utilities}
\alias{fillin_with_last_value}
\title{Fill-in missing trailing utilities with last value}
\usage{
fillin_trailing_utilities(utility, time_horizon)
}
\arguments{
\item{utility}{Vector of health quality of life values between 0 and 1}

\item{time_horizon}{Non-negative integer}
}
\value{
Vector of utilities
}
\description{
Repeat last value up to final period.
}
\seealso{
\code{\link{calc_QALY_population}}
}
