% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillin_utilities.R
\name{fillin_utilities}
\alias{fillin_utilities}
\title{fillin_utilities}
\usage{
fillin_utilities(utility, intervals = NA, time_horizon = NA)
}
\arguments{
\item{time_horizon}{}
}
\description{
intervals takes precedence over time_horizon
}
