% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor_error_conditions.R
\name{forbidden_negative_error}
\alias{forbidden_negative_error}
\title{forbidden_negative_error}
\usage{
forbidden_negative_error(subclass, message = "", call = sys.call(-1),
  ...)
}
\arguments{
\item{...}{}
}
\description{
forbidden_negative_error
}
\examples{
tryCatch(
  stop(forbidden_negative_error(message = "variable")),
  forbidden_negative_error = function(e) stop(e))
}
