% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-inflation.R
\docType{data}
\name{GDP_deflators}
\alias{GDP_deflators}
\title{ONS GDP_Deflators_Qtrly_National_Accounts}
\format{data.frame}
\source{
\url{https://www.gov.uk/government/uploads/system/uploads/attachment_data/file/562750/GDP_Deflators_Qtrly_National_Accounts_September_2016_update_v2.csv}
}
\usage{
GDP_deflators
}
\description{
ONS GDP_Deflators_Qtrly_National_Accounts
}
\keyword{datasets}
