% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSUV.R
\name{HSUV}
\alias{HSUV}
\title{Health State Utility Value}
\usage{
HSUV(..., method = c("prod", "add", "min"))
}
\arguments{
\item{...}{One or more utilities to combine}

\item{method}{Options are:
\itemize{
 \item Additive
\deqn{u^{add} = u_a + u_b - 1}
 \item Multiplicative
\deqn{u^{mult} = u_a u_b}
 \item Minimum
\deqn{u^{min} = \min{u_a, u_b}}
}}
}
\value{
function
}
\description{
Utilities can be combined in different ways to give an overall utility.
}
\details{
TODO: Adjusted decrement estimator
TODO: Combination model
}
\examples{

HSUV(1,1)
HSUV(1,1,2)

}
\references{
\url{http://nicedsu.org.uk/wp-content/uploads/2016/03/TSD12-Utilities-in-modelling-FINAL.pdf}
}
