% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discount.R
\name{make_discount}
\alias{make_discount}
\title{Make an Encapsulated Discount Function}
\usage{
make_discount(discount_rate = 0.035)
}
\arguments{
\item{discount_rate}{Discount factor, default at 3.5\%}
}
\value{
Function with global scoped index
}
\description{
This format doesn't need to keep track of time.
}
\examples{

}
