% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/person_health_years.R
\name{person_health_years}
\alias{person_health_years}
\title{Person health years object constructor}
\usage{
person_health_years(start_year = 0, end_year = NA, delay = 0,
  age = NA, time_horizon = NA, utility, discount_rate = 0.035,
  utility_method = "add")
}
\arguments{
\item{start_year}{Calendar year to begin calculation}

\item{end_year}{Calendar year to end calculation}

\item{delay}{If the start_year is in the future then the
first discounting is not 1 but smaller accordingly (positive integer)
We could alternatively define year_now}

\item{age}{Age at start of period (could be an age group)}

\item{time_horizon}{Number of time periods from start to end date}

\item{utility}{Proportion health detriment. This may change over time. Could provide as a list of length end-start or as say \code{list(c(2,3), c(u1,u2))} like in WinBUGs data.}

\item{discount_rate}{Fixed proportion reduction over time or a (continuous) function e.g. negative exponential}

\item{utility_method}{Should the yearly QALYs be summed to a scalar? \code{add} or \code{prod}}
}
\value{
Object of class person_health_years
\itemize{
 \item start_year
 \item end_year
 \item delay
 \item age
 \item time_horizon
 \item utility
 \item QoL
 \item discount_rate
 \item death
 \item utility_method
 }
}
\description{
For use as input to the QALY (and DALY) functions.
Want to provide set-up info in different ways but which gives us all the necessary information to go calc QALYs later.
individual level data.
Don’t need to supply all of these otherwise there would be redundancy and possible conflicts. E.g. time horizon can be got from start and end year.
Need to calc QoL which is discounting due to age from start to end year.
}
\examples{

personHealthYears <- person_health_years(
                      start_year = 2016,
                      end_year = 2020,
                      age = 33,
                      time_horizon = NA,
                      utility = 0.9,
                      discount_rate = 0.035)

total_QALYs(personHealthYears)
## 2.913622

\dontrun{
 total_QALY(1)
 ## "Error: Not an person_health_years class input object."
}
}
\seealso{
\code{\link{QALY}}, \code{\link{DALY}}
}
