% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.HRQoL}
\alias{plot.HRQoL}
\title{plot HRQoL_year}
\usage{
\method{plot}{HRQoL}(HRQoL_year, overlay = FALSE, XLIM = NA,
  COL = "light grey", age_annotate = FALSE)
}
\arguments{
\item{age_annotate}{}
}
\description{
plot HRQoL_year
}
