% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lexis.R
\name{plot_lexis.person_health_years}
\alias{plot_lexis.person_health_years}
\title{plot_lexis.person_health_years}
\usage{
plot_lexis.person_health_years(person_health_years)
}
\arguments{
\item{person_health_years}{}
}
\description{
##TODO: plot functions aren't finished
}
