% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QALY-package.R
\docType{package}
\name{QALY-package}
\alias{QALY}
\alias{QALY-package}
\title{QALY: Calculate Quality Adjusted Life Years (QALYs) Gains with Discounting and Inflated Costs}
\description{
S3 package for QALY calculation and plotting.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/n8thangreen/QALY}
}

}
\keyword{internal}
