% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QALY_survival.R
\name{QALY_survival}
\alias{QALY_survival}
\title{QALY_survival}
\usage{
QALY_survival(utility = NA, gender = NA, age = NA, start_delay = 0,
  discount_rate = 0.035, utility_method = "add")
}
\arguments{
\item{utility}{Vector of values between 0 and 1 (1 - utility loss)}

\item{age}{Year of age}

\item{start_delay}{What time delay to origin, to shift discounting}

\item{discount_rate}{default 3.5\%}

\item{utility_method}{Should the yearly QALYs be summed to a scalar? \code{add} or \code{prod}}

\item{intervals}{Time intervals for each utility}
}
\description{
TODO: use expected QALY over
(all-cause) life-table survival probabilities.
}
\details{
Uses the following formula:

\deqn{ QALY = \int_0^\infty S(t) u(t) dt \approx \sum_0^100 S(t) u(t) }
}
\examples{
calc_QALY(utility = 0.9,
          age = 13,
          intervals = 49)

}
\seealso{
\code{\link{calc_QALY_population}}
         \code{\link{total_QALYs}}
}
