% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_QALYs.R
\name{total_QALYs}
\alias{total_QALYs}
\alias{total_QALYs.default}
\alias{total_QALYs.person_health_years}
\title{Calculate Life-Time QALYs}
\usage{
total_QALYs(person_health_years)

\method{total_QALYs}{default}(person_health_years)

\method{total_QALYs}{person_health_years}(person_health_years)
}
\arguments{
\item{person_health_years}{Object of class \code{person_health_years}}
}
\value{
QALYs object
}
\description{
Calculate Life-Time QALYs
}
\examples{

personHealthYears <- person_health_years(
                      start_year = 2016,
                      end_year = 2020,
                      delay = 0,
                      age = NA,
                      time_horizon = NA,
                      utility = 0.9,
                      discount_rate = 0.035,
                      utility_method = "add")

total_QALYs(personHealthYears)
## 2.913622

\dontrun{
 total_QALYs(1)
 ## "Error: Not an person_health_years class input object."
}
}
