% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-inflation.R
\docType{data}
\name{annual_inflation_HCHS}
\alias{annual_inflation_HCHS}
\title{PSSRU annual inflation hospital and community health services}
\format{data.frame}
\source{
\url{https://www.pssru.ac.uk/pub/uc/uc2017/sources-of-information.pdf}
}
\usage{
annual_inflation_HCHS
}
\description{
PSSRU annual inflation hospital and community health services
}
\keyword{datasets}
